requirements.txt
openai-agents
requests
beautifulsoup4
python-dotenv
